@@ -1,35 +1,39 @@
 name: openemr
 description: A new Flutter project.
 
-version: 2.2+5
+version: 3.0.0
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
-  json_store: ^1.1.0
+  auth_buttons: ^3.0.2
+  barcode_widget: ^2.0.4
+  camera: ^0.10.5+4
+  charts_flutter_new: ^0.12.0
+  cloud_firestore: ^4.9.1
+  cupertino_icons: ^1.0.6
+  dash_chat: ^1.1.8
+  email_validator: ^2.1.17
+  firebase_auth: ^4.9.0
+  firebase_core: ^2.15.1
+  firebase_ml_vision: ^0.12.0+3
+  firebase_storage: ^11.2.6
   flutter:
     sdk: flutter
-  getwidget: ^1.2.4
-  webview_flutter: ^0.3.19+6
-  cupertino_icons: ^0.1.2
-  charts_flutter: ^0.9.0
-  wakelock: ^0.1.4+1
-  image_picker: ^0.6.7+22
-  dash_chat: ^1.1.8
-  camera: ^0.6.4+3
-  barcode_scan: ^3.0.1
-  shared_preferences: ^0.5.8
-  firebase_storage: ^3.1.6
-  cloud_firestore: ^0.13.7
-  firebase_auth: ^0.16.1
-  flutter_webrtc: ^0.2.8
-  google_sign_in: ^4.4.0
-  flutter_auth_buttons: ^0.8.0
-  firebase_ml_vision: ^0.9.6+2
-  http: ^0.12.2
-  modal_progress_hud: ^0.1.3
-  
+  flutter_lints: ^2.0.3
+  flutter_progress_hud: ^2.0.2
+  flutter_webrtc: ^0.9.41
+  getwidget: ^4.0.0
+  google_sign_in: ^6.1.5
+  http: ^1.1.0
+  image_picker: ^1.0.4
+  json_store: ^2.0.1
+  lints: ^2.1.1
+  shared_preferences: ^2.2.1
+  wakelock: ^0.6.2
+  webview_flutter: ^4.2.4
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
