@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:getwidget/getwidget.dart';
 import 'package:openemr/utils/rest_ds.dart';
@@ -7,15 +6,17 @@ import '../../models/user.dart';
 import './local_widgets/custom_dropdown_field.dart';
 
 class AddPatientScreen extends StatefulWidget {
+  const AddPatientScreen({super.key});
+
   @override
   _AddPatientScreenState createState() => _AddPatientScreenState();
 }
 
 class _AddPatientScreenState extends State<AddPatientScreen> {
-  User user;
+  late User user;
 
-  final formKey = new GlobalKey<FormState>();
-  String _title = "Mr.",
+  final formKey = GlobalKey<FormState>();
+  late String _title = "Mr.",
       _fname,
       _mname,
       _lname,
@@ -53,8 +54,7 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
   }
 
   void _showSnackBar(String text) {
-    ScaffoldMessenger.of(context)
-        .showSnackBar(new SnackBar(content: new Text(text)));
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(text)));
   }
 
   @override
@@ -68,7 +68,7 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
       },
       child: Scaffold(
         appBar: AppBar(
-          title: Text('Add Patient'),
+          title: const Text('Add Patient'),
           backgroundColor: Colors.black,
           centerTitle: true,
         ),
@@ -85,73 +85,73 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                   onStepCancel: cancel,
                   steps: [
                     Step(
-                      title: Text('Primary Info'),
+                      title: const Text('Primary Info'),
                       isActive: currentStep == 0,
                       content: Column(
                         children: [
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           CustomDropdownField(
                             label: 'Title',
-                            options: ['Mr.', 'Mrs.', 'Ms.', 'Dr.'],
+                            options: const ['Mr.', 'Mrs.', 'Ms.', 'Dr.'],
                             updateValue: (text) {
                               _title = text;
                             },
                             value: _title,
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter first name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _fname = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _fname = val!,
+                              decoration: const InputDecoration(
                                   border: OutlineInputBorder(),
                                   labelText: 'First Name'),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter middle name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _mname = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _mname = val!,
+                              decoration: const InputDecoration(
                                   border: OutlineInputBorder(),
                                   labelText: 'Middle Name'),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter middle name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _lname = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _lname = val!,
+                              decoration: const InputDecoration(
                                   border: OutlineInputBorder(),
                                   labelText: 'Last Name'),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
@@ -159,7 +159,7 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                               validator: (value) {
                                 final dobFormat = RegExp(
                                     r'[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]');
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter date of birth';
                                 }
                                 if (!dobFormat.hasMatch(value)) {
@@ -167,143 +167,143 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _dob = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _dob = val!,
+                              decoration: const InputDecoration(
                                   border: OutlineInputBorder(),
                                   labelText: 'DOB',
                                   hintText: "1999-08-09"),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           CustomDropdownField(
                             label: 'Gender',
-                            options: ['Male', 'Female'],
+                            options: const ['Male', 'Female'],
                             updateValue: (text) {
                               _sex = text;
                             },
                             value: _sex,
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                         ],
                       ),
                     ),
                     Step(
-                      title: Text('Address'),
+                      title: const Text('Address'),
                       isActive: currentStep == 1,
                       content: Column(
                         children: [
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter street name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _street = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _street = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Street',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter postal code';
                                 }
                                 return null;
                               },
                               keyboardType: TextInputType.number,
-                              onSaved: (val) => _postalCode = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _postalCode = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Postal Code',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter city name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _city = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _city = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'City',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter state name';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _state = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _state = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'State',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter country code';
                                 }
                                 return null;
                               },
                               keyboardType: TextInputType.number,
-                              onSaved: (val) => _countryCode = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _countryCode = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Country Code',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter phone number';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _phoneContact = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _phoneContact = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Phone Number',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                         ],
@@ -312,37 +312,37 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                     Step(
                       state: StepState.complete,
                       isActive: currentStep == 2,
-                      title: Text('Other Details'),
+                      title: const Text('Other Details'),
                       content: Column(
                         children: [
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter race';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _race = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _race = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Race',
                               ),
                             ),
                           ),
-                          SizedBox(
+                          const SizedBox(
                             height: 15,
                           ),
                           SizedBox(
                             child: TextFormField(
                               validator: (value) {
-                                if (value.isEmpty) {
+                                if (value!.isEmpty) {
                                   return 'Please enter ethnicity';
                                 }
                                 return null;
                               },
-                              onSaved: (val) => _ethnicity = val,
-                              decoration: InputDecoration(
+                              onSaved: (val) => _ethnicity = val!,
+                              decoration: const InputDecoration(
                                 border: OutlineInputBorder(),
                                 labelText: 'Ethnicity',
                               ),
@@ -354,7 +354,7 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                   ],
                 ),
               ),
-              SizedBox(
+              const SizedBox(
                 height: 15,
               ),
               GFButton(
@@ -362,7 +362,7 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
                 text: 'Add Patient',
                 color: GFColors.DARK,
               ),
-              SizedBox(
+              const SizedBox(
                 height: 15,
               ),
             ],
@@ -375,40 +375,40 @@ class _AddPatientScreenState extends State<AddPatientScreen> {
   void submit(context) async {
     final prefs = await SharedPreferences.getInstance();
     final form = formKey.currentState;
-    RestDatasource api = new RestDatasource();
-    if (form.validate()) {
-      form.save();
-      var baseUrl = prefs.getString('baseUrl');
-      var token = prefs.getString('token');
-      api
-          .addPatient(
-        baseUrl: baseUrl,
-        token: token,
-        title: _title.trim(),
-        fname: _fname.trim(),
-        mname: _mname.trim(),
-        lname: _lname.trim(),
-        dob: _dob.trim(),
-        sex: _sex.trim(),
-        street: _street.trim(),
-        postalcode: _postalCode.trim(),
-        city: _city.trim(),
-        state: _state.trim(),
-        countrycode: _countryCode.trim(),
-        phonecontact: _phoneContact.trim(),
-        race: _race.trim(),
-        ethnicity: _ethnicity.trim(),
-      )
-          .then(
-        (res) {
-          if (res != null && res["pid"] != null) {
-            form.reset();
-            _showSnackBar("Patient has been added");
-          } else {
-            _showSnackBar("Some error occured");
-          }
-        },
-      ).catchError((Object error) => _showSnackBar(error.toString()));
-    }
+    RestDatasource api = RestDatasource();
+    // if (form!.validate()) {
+    //   form.save();
+    //   var baseUrl = prefs.getString('baseUrl');
+    //   var token = prefs.getString('token');
+    //   api
+    //       .addPatient(
+    //     baseUrl: baseUrl,
+    //     token: token,
+    //     title: _title.trim(),
+    //     fname: _fname.trim(),
+    //     mname: _mname.trim(),
+    //     lname: _lname.trim(),
+    //     dob: _dob.trim(),
+    //     sex: _sex.trim(),
+    //     street: _street.trim(),
+    //     postalcode: _postalCode.trim(),
+    //     city: _city.trim(),
+    //     state: _state.trim(),
+    //     countrycode: _countryCode.trim(),
+    //     phonecontact: _phoneContact.trim(),
+    //     race: _race.trim(),
+    //     ethnicity: _ethnicity.trim(),
+    //   )
+    //       .then(
+    //     (res) {
+    //       if (res != null && res["pid"] != null) {
+    //         form.reset();
+    //         _showSnackBar("Patient has been added");
+    //       } else {
+    //         _showSnackBar("Some error occured");
+    //       }
+    //     },
+    //   ).catchError((Object error) => _showSnackBar(error.toString()));
   }
 }
+// }
