@@ -1,27 +1,29 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:getwidget/colors/gf_color.dart';
 import 'package:getwidget/components/button/gf_button.dart';
 import 'package:google_sign_in/google_sign_in.dart';
+import 'package:openemr/models/user.dart';
 import 'package:openemr/screens/telehealth/telehealth.dart';
 import 'package:openemr/utils/customlistloadingshimmer.dart';
 
 class CreateAccount extends StatefulWidget {
-  final FirebaseUser dispUser;
+  final User dispUser;
 
-  CreateAccount({Key key, @required this.dispUser}) : super(key: key);
+  const CreateAccount({required Key key, required this.dispUser})
+      : super(key: key);
   @override
   _CreateAccountBuyerState createState() => _CreateAccountBuyerState();
 }
 
 class _CreateAccountBuyerState extends State<CreateAccount> {
-  final userRef = Firestore.instance.collection('username');
+  final userRef = FirebaseFirestore.instance.collection('username');
   final GoogleSignIn googleSignIn = GoogleSignIn();
 
   bool _isLoading = false;
   final _formKey = GlobalKey<FormState>();
-  String _userid;
+  late String _userid;
 
   void _toggleLoadingStatus(bool newLoadingState) {
     setState(() {
@@ -32,17 +34,18 @@ class _CreateAccountBuyerState extends State<CreateAccount> {
   handleRegister(context) async {
     final form = _formKey.currentState;
 
-    if (form.validate()) {
+    if (form!.validate()) {
       form.save();
       _toggleLoadingStatus(true);
-      await userRef.document(widget.dispUser.uid).setData({
-        "id": _userid,
-        "name": widget.dispUser.displayName,
-      });
+      // await userRef.document(widget.dispUser.uid).setData({
+      //   "id": _userid,
+      //   "name": widget.dispUser.displayName,
+      // },
+      // );
 
       _toggleLoadingStatus(false);
       Navigator.of(context).pushReplacement(
-          MaterialPageRoute(builder: (context) => Telehealth()));
+          MaterialPageRoute(builder: (context) => const Telehealth()));
     }
   }
 
@@ -62,14 +65,14 @@ class _CreateAccountBuyerState extends State<CreateAccount> {
                   mainAxisAlignment: MainAxisAlignment.center,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: <Widget>[
-                    SizedBox(
+                    const SizedBox(
                       height: 25,
                     ),
                     Image.asset(
                       'lib/assets/images/firebase.png',
                       width: width * 0.25,
                     ),
-                    SizedBox(
+                    const SizedBox(
                       height: 20,
                     ),
                     _isLoading
@@ -80,18 +83,18 @@ class _CreateAccountBuyerState extends State<CreateAccount> {
                               SizedBox(
                                 child: TextFormField(
                                   validator: (value) {
-                                    if (value.isEmpty) {
+                                    if (value!.isEmpty) {
                                       return 'Username can\'t be blank';
                                     }
                                     return null;
                                   },
-                                  onSaved: (val) => _userid = val,
-                                  decoration: InputDecoration(
+                                  onSaved: (val) => _userid = val!,
+                                  decoration: const InputDecoration(
                                       border: OutlineInputBorder(),
                                       labelText: 'Username'),
                                 ),
                               ),
-                              SizedBox(
+                              const SizedBox(
                                 height: 20,
                               ),
                               GFButton(
