@@ -1,13 +1,15 @@
 import 'dart:async';
 
 import 'package:camera/camera.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:wakelock/wakelock.dart';
 import 'chart.dart';
 import 'package:getwidget/getwidget.dart';
-import 'package:flutter/cupertino.dart';
 
 class PPG extends StatefulWidget {
+  const PPG({super.key});
+
   @override
   PPGView createState() {
     return PPGView();
@@ -16,27 +18,27 @@ class PPG extends StatefulWidget {
 
 class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
   bool _toggled = false; // toggle button value
-  List<SensorValue> _data =
+  final List<SensorValue> _data =
       List<SensorValue>.empty(growable: true); // array to store the values
-  CameraController _controller;
-  double _alpha = 0.3; // factor for the mean value
-  AnimationController _animationController;
+  late CameraController _controller;
+  final double _alpha = 0.3; // factor for the mean value
+  late AnimationController _animationController;
   double _iconScale = 1;
   int _bpm = 0; // beats per minute
-  int _fs = 30; // sampling frequency (fps)
-  int _windowLen = 30 * 6; // window length to display - 6 seconds
-  CameraImage _image; // store the last camera image
-  double _avg; // store the average value during calculation
-  DateTime _now; // store the now Datetime
-  Timer _timer; // timer for image processing
+  final int _fs = 30; // sampling frequency (fps)
+  final int _windowLen = 30 * 6; // window length to display - 6 seconds
+  late CameraImage _image; // store the last camera image
+  late double _avg; // store the average value during calculation
+  late DateTime _now; // store the now Datetime
+  late Timer _timer; // timer for image processing
 
   @override
   void initState() {
     super.initState();
     _animationController =
-        AnimationController(duration: Duration(milliseconds: 500), vsync: this);
+        AnimationController(duration: const Duration(milliseconds: 500), vsync: this);
     _animationController
-      ..addListener(() {
+      .addListener(() {
         setState(() {
           _iconScale = 1.0 + _animationController.value * 0.4;
         });
@@ -45,12 +47,12 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
 
   @override
   void dispose() {
-    _timer?.cancel();
+    _timer.cancel();
     _toggled = false;
     _disposeController();
     Wakelock.disable();
-    _animationController?.stop();
-    _animationController?.dispose();
+    _animationController.stop();
+    _animationController.dispose();
     super.dispose();
   }
 
@@ -87,29 +89,29 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
                     Expanded(
                       flex: 1,
                       child: Padding(
-                        padding: EdgeInsets.all(12),
+                        padding: const EdgeInsets.all(12),
                         child: ClipRRect(
-                          borderRadius: BorderRadius.all(
+                          borderRadius: const BorderRadius.all(
                             Radius.circular(18),
                           ),
                           child: Stack(
                             fit: StackFit.expand,
                             alignment: Alignment.center,
                             children: <Widget>[
-                              _controller != null && _toggled
+                              _toggled
                                   ? AspectRatio(
                                       aspectRatio:
                                           _controller.value.aspectRatio,
                                       child: CameraPreview(_controller),
                                     )
                                   : Container(
-                                      padding: EdgeInsets.all(12),
+                                      padding: const EdgeInsets.all(12),
                                       alignment: Alignment.center,
                                       color: Colors.grey,
                                     ),
                               Container(
                                 alignment: Alignment.center,
-                                padding: EdgeInsets.all(4),
+                                padding: const EdgeInsets.all(4),
                                 child: Text(
                                   _toggled
                                       ? "Cover both the camera and the flash with your finger"
@@ -133,13 +135,13 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
                         mainAxisSize: MainAxisSize.min,
                         crossAxisAlignment: CrossAxisAlignment.center,
                         children: <Widget>[
-                          Text(
+                          const Text(
                             "Estimated BPM",
                             style: TextStyle(fontSize: 18, color: Colors.grey),
                           ),
                           Text(
                             (_bpm > 30 && _bpm < 150 ? _bpm.toString() : "--"),
-                            style: TextStyle(
+                            style: const TextStyle(
                                 fontSize: 32, fontWeight: FontWeight.bold),
                           ),
                         ],
@@ -171,8 +173,8 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
             Expanded(
               flex: 1,
               child: Container(
-                margin: EdgeInsets.all(12),
-                decoration: BoxDecoration(
+                margin: const EdgeInsets.all(12),
+                decoration: const BoxDecoration(
                     borderRadius: BorderRadius.all(
                       Radius.circular(18),
                     ),
@@ -190,18 +192,19 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
     // create array of 128 ~= 255/2
     _data.clear();
     int now = DateTime.now().millisecondsSinceEpoch;
-    for (int i = 0; i < _windowLen; i++)
+    for (int i = 0; i < _windowLen; i++) {
       _data.insert(
           0,
           SensorValue(
               DateTime.fromMillisecondsSinceEpoch(now - i * 1000 ~/ _fs), 128));
+    }
   }
 
   void _toggle() {
     _clearData();
     _initController().then((onValue) {
       Wakelock.enable();
-      _animationController?.repeat(reverse: true);
+      _animationController.repeat(reverse: true);
       setState(() {
         _toggled = true;
       });
@@ -214,38 +217,38 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
   void _untoggle() {
     _disposeController();
     Wakelock.disable();
-    _animationController?.stop();
-    _animationController?.value = 0.0;
+    _animationController.stop();
+    _animationController.value = 0.0;
     setState(() {
       _toggled = false;
     });
   }
 
   void _disposeController() {
-    _controller?.dispose();
-    _controller = null;
+    _controller.dispose();
+    _controller;
   }
 
   Future<void> _initController() async {
     try {
-      List _cameras = await availableCameras();
-      _controller = CameraController(_cameras.first, ResolutionPreset.low);
+      List cameras = await availableCameras();
+      _controller = CameraController(cameras.first, ResolutionPreset.low);
       await _controller.initialize();
-      Future.delayed(Duration(milliseconds: 100)).then((onValue) {
+      Future.delayed(const Duration(milliseconds: 100)).then((onValue) {
         _controller.setFlashMode(FlashMode.torch);
       });
       _controller.startImageStream((CameraImage image) {
         _image = image;
       });
     } catch (Exception) {
-      debugPrint(Exception);
+      debugPrint(Exception as String?);
     }
   }
 
   void _initTimer() {
     _timer = Timer.periodic(Duration(milliseconds: 1000 ~/ _fs), (timer) {
       if (_toggled) {
-        if (_image != null) _scanImage(_image);
+        _scanImage(_image);
       } else {
         timer.cancel();
       }
@@ -272,44 +275,44 @@ class PPGView extends State<PPG> with SingleTickerProviderStateMixin {
     // Since this function doesn't need to be so "exact" regarding the time it executes,
     // I only used the a Future.delay to repeat it from time to time.
     // Ofc you can also use a Timer object to time the callback of this function
-    List<SensorValue> _values;
-    double _avg;
-    int _n;
-    double _m;
-    double _threshold;
-    double _bpm;
-    int _counter;
-    int _previous;
+    List<SensorValue> values;
+    double avg;
+    int n;
+    double m;
+    double threshold;
+    double bpm;
+    int counter;
+    int previous;
     while (_toggled) {
-      _values = List.from(_data); // create a copy of the current data array
-      _avg = 0;
-      _n = _values.length;
-      _m = 0;
-      _values.forEach((SensorValue value) {
-        _avg += value.value / _n;
-        if (value.value > _m) _m = value.value;
-      });
-      _threshold = (_m + _avg) / 2;
-      _bpm = 0;
-      _counter = 0;
-      _previous = 0;
-      for (int i = 1; i < _n; i++) {
-        if (_values[i - 1].value < _threshold &&
-            _values[i].value > _threshold) {
-          if (_previous != 0) {
-            _counter++;
-            _bpm += 60 *
+      values = List.from(_data); // create a copy of the current data array
+      avg = 0;
+      n = values.length;
+      m = 0;
+      for (var value in values) {
+        avg += value.value / n;
+        if (value.value > m) m = value.value;
+      }
+      threshold = (m + avg) / 2;
+      bpm = 0;
+      counter = 0;
+      previous = 0;
+      for (int i = 1; i < n; i++) {
+        if (values[i - 1].value < threshold &&
+            values[i].value > threshold) {
+          if (previous != 0) {
+            counter++;
+            bpm += 60 *
                 1000 /
-                (_values[i].time.millisecondsSinceEpoch - _previous);
+                (values[i].time.millisecondsSinceEpoch - previous);
           }
-          _previous = _values[i].time.millisecondsSinceEpoch;
+          previous = values[i].time.millisecondsSinceEpoch;
         }
       }
-      if (_counter > 0) {
-        _bpm = _bpm / _counter;
-        print(_bpm);
+      if (counter > 0) {
+        bpm = bpm / counter;
+        print(bpm);
         setState(() {
-          this._bpm = ((1 - _alpha) * _bpm + _alpha * _bpm).toInt();
+          _bpm = ((1 - _alpha) * bpm + _alpha * bpm).toInt();
         });
       }
       await Future.delayed(Duration(
