@@ -3,12 +3,13 @@ import 'package:flutter/material.dart';
 class SystemPadding extends StatelessWidget {
   final Widget child;
 
-  SystemPadding({Key key, this.child}) : super(key: key);
+  const SystemPadding({required Key key, required this.child})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     var mediaQuery = MediaQuery.of(context);
-    return new AnimatedContainer(
+    return AnimatedContainer(
         padding: mediaQuery.viewInsets,
         duration: const Duration(milliseconds: 300),
         child: child);
